+++ CHANGES.md
@@ -9,6 +9,8 @@
 -   Added BAP recipe for standard cwe_checker run (PR #9)
 -   Improved check for CWE-476 (NULL Pointer Dereference) using data flow analysis (PR #11)
 -   Switched C build system from make to scons (PR #16)
+-   Added type inference pass (PR #14)
+-   Added unit tests to test suite (PR #14)
 
 0.1 (2018-10-08)
 =====